requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.24.0
sqlalchemy==2.0.41
alembic==1.13.1
pillow==11.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
structlog==23.2.0
fastapi-limiter==0.1.5
websockets==12.0
pydantic==2.10.4
pydantic-settings==2.2.1
email-validator==2.1.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
